refactor: rename project to projects
